Automate table creation and data population with populate-database.js
