--- db/seeding/header-line.tsv
@@ -1,1 +1,0 @@
-reviewId	date	helpful	images 	name	pageId	profilepicture	stars	text	title	username	verified
